Allow to use webdav storage as cache storage
